Initial package, restructuring
