+++ README.md
@@ -1,3 +1,14 @@
 # SWAMP
 
 NOAA-ARL/ATDD Soil Water Analysis Model Product (SWAMP)
+
+## Development installation
+
+1. Create Conda environment based on `environment-dev.yml`.
+   ```
+   conda env create -f environment-dev.yml
+   ```
+2. Install `swamp` package using
+   ```
+   pip install -e . --no-deps
+   ```
