+++ setup.cfg
@@ -1,3 +1,29 @@
+[metadata]
+name = swamp
+version = attr: swamp.__version__
+description = NOAA-ARL/ATDD Soil Water Analysis Model Product (SWAMP)
+long_description = file: README.md
+url = https://github.com/noaa-oar-arl/SWAMP
+license = MIT
+author = Zachary Moon
+author_email = zachary.moon@noaa.gov
+maintainer = Zachary Moon
+maintainer_email = zachary.moon@noaa.gov
+
+[options]
+zip_safe = False
+packages = find:
+include_package_data = True
+python_requires = >=3.7
+install_requires =
+  cartopy
+  matplotlib
+  metpy
+  numpy
+  pandas
+  requests
+  xarray
+
 [flake8]
 max-line-length = 100
 ignore =
