+++ swamp/__init__.py
@@ -1,0 +1,6 @@
+"""
+NOAA-ARL/ATDD Soil Water Analysis Model Product (SWAMP)
+"""
+__version__ = "0.1.0.dev0"
+
+from .load import get_alexi, get_crn, get_prism, load_alexi  # noqa: F401
